+++ Sources/KeyboardKitSwiftUI/_Deprecated/View+Deprecated.swift
@@ -1,0 +1,10 @@
+import SwiftUI
+
+public extension View {
+    
+    @available(*, deprecated, message: "This modifier is no longer used by KeyboardKit and will be removed in 4.0.")
+    func systemKeyboardButtonFrame(for style: SystemKeyboardStyle)  -> some View {
+        self.frame(maxWidth: .infinity)
+            .frame(height: style.buttonHeight)
+    }
+}